a very mysterious method added
the method checkAndCreateTrades ** somehow ** works. I dont know why, but I do know that if add a simple iterator item remove on a COPIED list it stops working. Guess thats recursion some some other weird thing. Point is, it is working, I dont ever want to change it even if my grade depends on it. This method is trash and perfect in every way possible. Today it does work, but we shall see what tomorrow brings.